requirements.txt
langgraph>=0.0.30
pysimular>=0.1.0
playwright>=1.41.0
langchain>=0.1.0
langchain-community>=0.0.16
langchain-openai>=0.0.3
langchain-anthropic>=0.0.3
pypdf>=3.17.0
click>=8.1.7
python-dotenv>=1.0.0
anthropic>=0.8.0
openai>=1.6.0
typer>=0.9.0
rich>=13.7.0
tavily-python>=0.2.6
